backend/requirements.txt
azure-ai-inference
azure-identity
azure-cognitiveservices-speech
fastapi
uvicorn
python-dotenv
Jinja2
